Consider column filter when choosing multival cols
Users will not see blacklisted columns (or non-whitelisted,
if there is a whitelist) when choosing which columns are
multivalued.

Also fixed a bug where I accidentally modified the
'columns' list in filter_columns() and
ask_multivalue_columns(). By modifying it in the first function,
I got an incomplete list in the second.